Spongebob death done - I am so close to finishing this game to a playable state.

+++ src/UI/GameMenu.java
@@ -14,6 +14,7 @@ import engineTester.Clock;
 import entities.Entity;
 import entities.Typer;
 import shapes.Shapes;
+import turnController.MasterTurnController;
 
 public class GameMenu extends Entity {
 
@@ -21,23 +22,25 @@ public class GameMenu extends Entity {
 	public int selected;
 
 	public float creditsY;
-	
+
 	public float delay, timeSinceLastPress, timer;
 
 	// Audio
 	int genericMenuSelectBuffer, MenuMusicBuffer, sndNoiseBuffer, creditsBuffer;
 	public Source source, MenuMusicSource, sndNoiseSource, creditsSource;
-	private mode mode;
+	public mode mode;
 
 	private Texture titleTexture, heartTexture, creditsTexture;
 
 	private Typer typer;
-	
+
 	enum mode {
 		MENU, PLAY, TRANSISTION, CLOSE, CREDITS
 	}
 
-	public GameMenu(float x, float y) {
+	public MasterTurnController turnController;
+	
+	public GameMenu(float x, float y, MasterTurnController turnController) {
 		super(x, y);
 		this.selected = 0;
 		this.menu_options = new ArrayList<Button>();
@@ -67,37 +70,37 @@ public class GameMenu extends Entity {
 		this.heartTexture = Shapes.LoadTexture("res/heart.png", "PNG");
 
 		this.timer = 10f;
-		
+
 		this.creditsTexture = Shapes.LoadTexture("res/credits.png", "PNG");
 		this.creditsY = 900;
-		
+
 		this.creditsBuffer = AudioMaster.loadSound("audio/Fin.wav");
 		this.creditsSource = new Source();
 		this.typer = new Typer(25, 900, 18, "Press Escape to return back to main menu");
 		this.typer.setCurrentText("Press Escape to return back to main menu");
 		this.typer.setRenderStar(false);
-		
+
+		this.turnController = turnController;
 	}
 
 	@Override
 	public void render() {
-		
-		// black background 
+
+		// black background
 		GL11.glColor3f(0, 0, 0);
 		Shapes.draw_quad(0, 0, 1280, 960);
-		
+
 		if (this.mode == mode.MENU) {
 			for (Entity button : menu_options) {
 				button.render();
 			}
 			GL11.glEnable(GL_TEXTURE_2D);
 			GL11.glColor3f(0, 0, 1);
-			Shapes.DrawQuadTex(heartTexture, 520, 332 + selected*200, 32, 32);
+			Shapes.DrawQuadTex(heartTexture, 520, 332 + selected * 200, 32, 32);
 
 			// Spongetale title page
 			GL11.glColor3f(1, 1, 1);
 			Shapes.DrawQuadTex(titleTexture, 290, 0, 700, 300);
-			
 
 		} else if (this.mode == mode.CREDITS) {
 			GL11.glEnable(GL_TEXTURE_2D);
@@ -105,7 +108,7 @@ public class GameMenu extends Entity {
 			Shapes.DrawQuadTex(creditsTexture, 0, creditsY, 1280, 3840);
 			creditsY -= 1;
 			typer.render();
-		}else {
+		} else {
 			GL11.glColor3f(0, 0, 0);
 			Shapes.draw_quad(0, 0, 1280, 960);
 		}
@@ -128,30 +131,32 @@ public class GameMenu extends Entity {
 			}
 
 			if (Keyboard.isKeyDown(Keyboard.KEY_RETURN)) {
-				if (selected == 0) {
-					this.mode = mode.TRANSISTION;
-					MenuMusicSource.stop();
-					sndNoiseSource.play(sndNoiseBuffer);
-				} else if (selected == 1){
-					MenuMusicSource.stop();
-					this.mode = mode.TRANSISTION;
-					sndNoiseSource.play(sndNoiseBuffer);
-				} else if (selected == 2) {
-					MenuMusicSource.delete();
-					this.mode = mode.CLOSE;
+				if (timeSinceLastPress > delay) {
+					if (selected == 0) {
+						this.mode = mode.TRANSISTION;
+						MenuMusicSource.stop();
+						sndNoiseSource.play(sndNoiseBuffer);
+					} else if (selected == 1) {
+						MenuMusicSource.stop();
+						this.mode = mode.TRANSISTION;
+						sndNoiseSource.play(sndNoiseBuffer);
+					} else if (selected == 2) {
+						MenuMusicSource.delete();
+						this.mode = mode.CLOSE;
+					}
 				}
-
 			}
 
 			timeSinceLastPress += Clock.Delta();
 		} else {
-			if (timer > 0){
+			if (timer > 0) {
 				timer -= Clock.Delta();
 			} else {
-				if (selected == 0){
+				if (selected == 0) {
+					turnController.reset();
 					this.mode = mode.PLAY;
 					sndNoiseSource.play(sndNoiseBuffer);
-				} else if (selected == 1){
+				} else if (selected == 1) {
 					this.mode = mode.CREDITS;
 					creditsSource.play(creditsBuffer);
 					creditsSource.setVolume(0.4f);
@@ -173,15 +178,15 @@ public class GameMenu extends Entity {
 			timeSinceLastPress = 0;
 			source.play(genericMenuSelectBuffer);
 
-			if (selected == 0){
+			if (selected == 0) {
 				menu_options.get(0).setTexture("play bar selected");
 				menu_options.get(1).setTexture("credits bar");
 				menu_options.get(2).setTexture("exit bar");
-			} else if (selected == 1){
+			} else if (selected == 1) {
 				menu_options.get(0).setTexture("play bar");
 				menu_options.get(1).setTexture("credits bar selected");
 				menu_options.get(2).setTexture("exit bar");
-			} else if (selected == 2){
+			} else if (selected == 2) {
 				menu_options.get(0).setTexture("play bar");
 				menu_options.get(1).setTexture("credits bar");
 				menu_options.get(2).setTexture("exit bar selected");
@@ -197,4 +202,25 @@ public class GameMenu extends Entity {
 	public Boolean playerRequestToClose() {
 		return this.mode == mode.CLOSE;
 	}
+
+	public mode getMode() {
+		return mode;
+	}
+
+	public void setMode(mode mode) {
+		this.mode = mode;
+	}
+
+	public void close() {
+		this.mode = mode.CLOSE;
+	}
+
+	public void mainMenu() {
+		this.mode = mode.MENU;
+	}
+
+	public void tempFreeze() {
+		timeSinceLastPress = -8f;
+	}
+
 }
